fix: fix pytest related failure during migration apply
